# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IT Interview Prep API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaderboard:
    get:
      operationId: listLeaderboard
      tags: [leaderboard]
      summary: Get top scorers leaderboard
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: level
          in: query
          required: false
          schema:
            type: ["string", "null"]
            enum: [beginner, intermediate, advanced]
        - name: timedMode
          in: query
          required: false
          schema:
            type: ["boolean", "null"]
          description: Filter by timed mode (true/false/null for all)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Leaderboard entries ordered by score desc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
    post:
      operationId: createLeaderboardEntry
      tags: [leaderboard]
      summary: Save a player score to the leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaderboardEntryInput"
      responses:
        "201":
          description: Entry saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardEntry"
...
  /me/scores:
    get:
      tags: [leaderboard]
      operationId: listMyScores
      summary: Get the authenticated user's personal score history
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: The user's leaderboard entries, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LeaderboardEntry:
      type: object
      required: [id, playerName, courseId, courseName, level, score, totalQuestions, percentage, badges, createdAt]
      properties:
        id:
          type: integer
        playerName:
          type: string
        userId:
          type: ["string", "null"]
          description: ID of the authenticated user who posted this score, if any
        courseId:
          type: integer
        courseName:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        score:
          type: integer
        totalQuestions:
          type: integer
        percentage:
          type: number
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        timedMode:
          type: boolean
          description: Whether this score was achieved in timed mode
        timeBonus:
          type: integer
          description: Time bonus points earned
    LeaderboardEntryInput:
      type: object
      required: [sessionId, playerName]
      properties:
        sessionId:
          type: integer
          description: The session ID whose server-computed result should be recorded
        playerName:
          type: string
...